css/css-scroll-snap-2/scroll-start-with-programmatic-scroll.html

<!DOCTYPE html>
<html>

<head>
  <meta charset="utf-8">
  <title> CSS Scroll Snap 2 Test: scroll-start-*</title>
  <link rel="help" href="https://drafts.csswg.org/css-scroll-snap-2/#scroll-start">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<style>
  #scroller {
    height: 100px;
    width: 100px;
    overflow: scroll;
    scroll-start: 10vh 200px;
  }

  .spacer {
    width: 200vw;
    height: 200vh;
    border: solid 1px green;
  }
</style>
<script>
  function runTest() {
        // scroll position declared by scroll-start.
    const scroll_start_top = 0.1 * window.innerHeight;
    const scroll_start_left = 200;

    // target position of the programmatic scroll.
    const target_scroll_top = 100;
    const target_scroll_left = 100;

    promise_test(async (t) => {
      // verify that we are starting from the offsets indicated by scroll start.
      assert_equals(scroller.scrollTop, scroll_start_top,
        "scroll-start: <length> sets initial scroll position vertically");
      assert_equals(scroller.scrollLeft, scroll_start_left,
        "scroll-start: <length> sets initial scroll position horizontally");

      // verify that the programmatic scroll should result in an actual scroll.
      assert_not_equals(target_scroll_top, scroll_start_top,
        "programmatic scroll should not be a nop vertically");
      assert_not_equals(target_scroll_left, scroll_start_left,
        "programmatic scroll should not be a nop horizontally");

      scroller.scrollTop = target_scroll_top;
      scroller.scrollLeft = target_scroll_left;
      // we might need to wait a frame for the programmatic scroll to take
      // effect.
      await requestAnimationFrame(() => {});

      // verify that programmtic scroll succeeded.
      assert_equals(scroller.scrollTop, target_scroll_top,
        "programmatic scroll succeeds vertically");
      assert_equals(scroller.scrollLeft, target_scroll_left,
        "programmatic scroll succeeds horizontally");
    });
  }
</script>

<body onload="runTest()">
  <div id="scroller">
    <div class="spacer"></div>
  </div>
</body>
